remove old 4.3->5.0 bbb controlpanel adapter.
