--- Products/CMFPlone/controlpanel/bbb/configure.zcml
@@ -1,18 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:five="http://namespaces.zope.org/five"
-    xmlns:browser="http://namespaces.zope.org/browser">
-
-  <adapter factory=".editing.EditingControlPanelAdapter" />
-  <adapter factory=".language.LanguageControlPanelAdapter" />
-  <adapter factory=".filter.FilterControlPanelAdapter" />
-  <adapter factory=".maintenance.MaintenanceControlPanelAdapter" />
-  <adapter factory=".mail.MailControlPanelAdapter" />
-  <adapter factory=".navigation.NavigationControlPanelAdapter" />
-  <adapter factory=".search.SearchControlPanelAdapter" />
-  <adapter factory=".security.SecurityControlPanelAdapter" />
-  <adapter factory=".site.SiteControlPanelAdapter" />
-  <adapter factory=".markup.MarkupControlPanelAdapter" />
-  <adapter factory=".usergroups.UserGroupsSettingsControlPanelAdapter" />
-
-</configure>
