+++ Products/CMFPlone/controlpanel/browser/maintenance.py
@@ -3,34 +3,42 @@ from AccessControl.Permissions import view_management_screens
 from Acquisition import aq_inner
 from App.config import getConfiguration
 from plone.autoform.form import AutoExtensibleForm
+from plone.registry.interfaces import IRegistry
+from plone.base.interfaces import IMaintenanceSchema
+from plone.base.utils import human_readable_size
 from plone.memoize.view import memoize
 from plone.protect import CheckAuthenticator
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as _
-from plone.base.interfaces import IMaintenanceSchema
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from z3c.form import button
 from z3c.form import form
 from zope.component import getMultiAdapter
+from zope.component import getUtility
+from zope.interface import implementer
 
 import logging
-import os
 import time
-
-LIFETIME = True
-try:
-    from Lifetime import shutdown
-except ImportError:
-    LIFETIME = False
-
-try:
-    from html import escape
-except ImportError:
-    from cgi import escape
 
 
 logger = logging.getLogger(__file__)
 
+@implementer(IMaintenanceSchema)
+class MaintenanceControlPanelAdapter:
+
+    def __init__(self, context):
+        self.context = context
+        registry = getUtility(IRegistry)
+        self.maintenance_settings = registry.forInterface(
+            IMaintenanceSchema, prefix="plone")
+
+    def get_days(self):
+        return self.maintenance_settings.days
+
+    def set_days(self, value):
+        self.maintenance_settings.days = value
+
+    days = property(get_days, set_days)
 
 class MaintenanceControlPanel(AutoExtensibleForm, form.EditForm):
     """A simple form to pack the databases."""
@@ -72,67 +80,11 @@ class MaintenanceControlPanel(AutoExtensibleForm, form.EditForm):
             db.pack(t)
         self.status = _('Packed the database.')
 
-    @button.buttonAndHandler(_('Shut down'), name='shutdown')
-    def handle_shutdown_action(self, action):
-        CheckAuthenticator(self.request)
-        if not self.available():
-            self.status = _(
-                'text_not_allowed_manage_server',
-                default='You are not allowed to manage the Zope server.'
-            )
-            return
-        try:
-            user = '"%s"' % getSecurityManager().getUser().getUserName()
-        except:
-            user = 'unknown user'
-        logger.info("Shutdown requested by %s" % user)
-        if LIFETIME:
-            shutdown(0)
-        else:
-            raise
-        # TODO: returning html has no effect in button handlers
-        self.request.response.setHeader('X-Theme-Disabled', 'True')
-        return """<html><head></head><body>{}</body></html>""".format(
-            _('plone_shutdown', default="Zope is shutting down.")
-        )
-
-    @button.buttonAndHandler(_('Restart'), name='restart')
-    def handle_restart_action(self, action):
-        CheckAuthenticator(self.request)
-        if not self.available():
-            self.status = _(
-                'text_not_allowed_manage_server',
-                default='You are not allowed to manage the Zope server.'
-            )
-            return
-
-        try:
-            user = '"%s"' % getSecurityManager().getUser().getUserName()
-        except:
-            user = 'unknown user'
-        logger.info("Restart requested by %s" % user)
-        shutdown(1)
-        url = self.request.get('URL')
-        # TODO: returning html has no effect in button handlers
-        self.request.response.setHeader('X-Theme-Disabled', 'True')
-        return """<html><head>
-            <meta http-equiv="refresh" content="5; url={}">
-        </head><body>{}</body></html>""".format(
-            escape(url, 1),
-            _('plone_restarting',
-                default="Zope is restarting. This page will refresh in 30"
-                        " seconds...")
-        )
-
     def available(self):
         root = aq_inner(self.context).getPhysicalRoot()
         sm = getSecurityManager()
         return sm.checkPermission(view_management_screens, root)
 
-    def isRestartable(self):
-        if 'ZMANAGED' in os.environ:
-            return True
-        return False
 
     def isDevelopmentMode(self):
         return bool(getConfiguration().debug_mode)
@@ -148,11 +100,4 @@ class MaintenanceControlPanel(AutoExtensibleForm, form.EditForm):
 
     def dbSize(self):
         size = self.portal()._p_jar.db().getSize()
-
-        # From Zope2.App.ApplicationManager
-        if type(size) is str:
-            return size
-
-        if size >= 1048576.0:
-            return '%.1f MB' % (size / 1048576.0)
-        return '%.1f kB' % (size / 1024.0)
+        return human_readable_size(size)
