+++ Products/CMFPlone/controlpanel/configure.zcml
@@ -4,7 +4,6 @@
 
   <include package="plone.app.registry" />
 
-  <include package=".bbb" />
   <include package=".browser" />
 
   <include file="events.zcml" />
