+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_browser_maintenance.py
@@ -1,13 +1,18 @@
-from Products.CMFPlone.testing import PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
+from pkg_resources import get_distribution
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 from plone.app.testing import TEST_USER_NAME
 from plone.app.testing import TEST_USER_PASSWORD
+from plone.base.interfaces import IMaintenanceSchema
+from plone.registry.interfaces import IRegistry
 from plone.testing.zope import Browser
 from plone.testing.zope import login
-import unittest
-from App.ApplicationManager import ApplicationManager
-from pkg_resources import get_distribution
+from Products.CMFPlone.testing import PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
+from Products.CMFPlone.testing import PRODUCTS_CMFPLONE_INTEGRATION_TESTING
+from zope.component import getAdapter
+from zope.component import getUtility
 
-has_zope4 = get_distribution('Zope2').version.startswith('4')
+import unittest
 
 
 class MaintenanceControlPanelFunctionalTest(unittest.TestCase):
@@ -86,3 +91,43 @@ class MaintenanceControlPanelFunctionalTest(unittest.TestCase):
         self.assertTrue(self.browser.url.endswith('maintenance-controlpanel'))
         self.assertTrue('Packed the database.' in self.browser.contents)
         db.pack = original_pack
+
+
+class MaintenanceControlPanelAdapterTest(unittest.TestCase):
+
+    layer = PRODUCTS_CMFPLONE_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        registry = getUtility(IRegistry)
+        self.maintenance_settings = registry.forInterface(
+            IMaintenanceSchema, prefix="plone")
+
+    def test_adapter_lookup(self):
+        self.assertTrue(
+            getAdapter(self.portal, IMaintenanceSchema)
+        )
+
+    def test_get_days(self):
+        self.assertEqual(
+            getAdapter(self.portal, IMaintenanceSchema).days,
+            7
+        )
+        self.maintenance_settings.days = 4
+        self.assertEqual(
+            getAdapter(self.portal, IMaintenanceSchema).days,
+            4
+        )
+
+    def test_set_days(self):
+        self.assertEqual(
+            self.maintenance_settings.days,
+            7
+        )
+        getAdapter(self.portal, IMaintenanceSchema).days = 4
+        self.assertEqual(
+            self.maintenance_settings.days,
+            4
+        )
