+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_browser_usergroups.py
@@ -1,10 +1,8 @@
 from plone.app.testing import SITE_OWNER_NAME, SITE_OWNER_PASSWORD
-from Products.CMFCore.utils import getToolByName
 from plone.testing.zope import Browser
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.testing import  PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
 from Products.CMFPlone.utils import normalizeString
-
-from Products.CMFPlone.testing import \
-    PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
 
 import unittest
 import transaction
