+++ news/3494.breaking
@@ -1,0 +1,2 @@
+Remove old 4.3->5.0 bbb  controlpanel adapter.
+[jensens]
